01 11/10/2023

+++ src/app/cart/cart.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ProductService} from "../services/product.service";
-import {Cart} from "../model/product.model";
+import {Cart, Product} from "../model/product.model";
+import {Observable} from "rxjs";
 
 @Component({
   selector: 'app-cart',
@@ -11,25 +12,18 @@ export class CartComponent implements OnInit{
   constructor(private productService : ProductService) {
   }
 
-  products : Array<Cart> = []
+  products$! : Observable<Array<Cart>>
 
   ngOnInit(): void {
     this.getProductsFromCart()
   }
 
   getProductsFromCart(){
-    this.productService.getProductsFromCart()
-      .subscribe({
-        next : products => {
-          this.products = products
-        },
-        error : err => {
-          alert("Errrrorrr!!!!!")
-        }
-      })
+    this.products$ = this.productService.getProductsFromCart()
   }
 
   removeFromCart(product: Cart) {
+    if(confirm("Do really want to delete this product"))
     this.productService.removeFromCart(product)
       .subscribe({
         next : value => {
